Moving `prepareIndexRequest` to separate, implementation class.
